+++ packages/drizzle/migrations/20260511012330_married_moonstone/migration.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "submissions"."submission_location_photo_selections" ADD COLUMN "is_main" boolean DEFAULT false NOT NULL;